Use setuptools_scm to get the version number
Follow recommendation from https://jazzband.co/about/releases#packaging.

Files tracked with the version control system are included in the source
distribution. That avoids the error-prone listing of files in
MANIFEST.in.

Don’t package the .github directory and sonar-project.properties, they
contain CI configuration for the project and should not be distributed.

Don’t package .gitignore, it is specific to version control and should
not be distributed.

Refs #224

Reviewed-by: Stefan Foulis