Allow sorting PhoneNumber instances in Python
While the use case is not obvious, databases allow sorting by a phone
number field. Users can reasonably expect to sort that data in Python as
well.

The __lt__ method expects an homogeneous collection of PhoneNumber
instances, or possibly `phonenumbers.PhoneNumber`.

The database string representation of PhoneNumber is used for the
comparison, to reasonably match the sorting that the database would
perform. It may differ based on the database collation.

Fix #371

Reviewed-by: Stefan Foulis